requirements.txt
pandas>=1.3.0
neo4j>=5.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
tqdm>=4.62.0
numpy>=1.21.0 